// indiakart_homepage/src/app/Model/RetailerRegister.ts
export interface RetailerRegister {
    id: number;
    firstName: string;
    lastName: string;
    phone: string;
    email: string;
    isActive: boolean;
    companyName:string;
    companyPhone:string;
    companyAddress:string;
    username: string;
    password: string;
    lastLogin: string;
}